Updates per Toby
Remove mine step
Remove extra data file
Clean up JSON formatting
Move 5D definition into 18CO file
Remove unused functions
